# backend/tests/snapshots/openapi.json
{
 "components": {
  "schemas": {
...
   "Badcase": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "number"
     },
     "expected": {
      "default": "",
      "title": "Expected",
      "type": "string"
     },
     "fix_commit": {
      "default": "",
      "title": "Fix Commit",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "input": {
      "default": "",
      "title": "Input",
      "type": "string"
     },
     "item_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Item Id"
     },
     "problem": {
      "title": "Problem",
      "type": "string"
     },
     "regression_case": {
      "default": "",
      "title": "Regression Case",
      "type": "string"
     },
     "root_cause": {
      "enum": [
       "U",
       "P",
       "R",
       "W",
       "V",
       "B",
       "E",
       "X",
       "S",
       "K",
       "UX"
      ],
      "title": "Root Cause",
      "type": "string"
     },
     "run_id": {
      "title": "Run Id",
      "type": "string"
     },
     "severity": {
      "default": "S2",
      "enum": [
       "S1",
       "S2",
       "S3"
      ],
      "title": "Severity",
      "type": "string"
     },
     "stage": {
      "default": "",
      "title": "Stage",
      "type": "string"
     },
     "status": {
      "default": "open",
      "enum": [
       "open",
       "fixed",
       "wontfix"
      ],
      "title": "Status",
      "type": "string"
     },
     "tags": {
      "items": {
       "type": "string"
      },
      "title": "Tags",
      "type": "array"
     }
    },
    "required": [
     "run_id",
     "problem",
     "root_cause",
     "id",
     "created_at"
    ],
    "title": "Badcase",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "VeriChalk",
  "version": "0.1.0.dev0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/debug/badcases": {
   "get": {
    "operationId": "debug_badcases_api_debug_badcases_get",
    "parameters": [
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Badcase"
         },
         "title": "Response Debug Badcases Api Debug Badcases Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Badcases"
   },
   "post": {
    "description": "把一次运行里发现的坏例子写入 Badcase 簿（YAML 文件，之后按根因类别修复并转成回归用例）。",
    "operationId": "debug_add_badcase_api_debug_badcases_post",
    "parameters": [
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BadcaseIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Badcase"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Add Badcase"
   }
  },
  "/api/debug/kp/{kp_id}": {
   "get": {
    "description": "图检索视图里点击节点：知识点说明、掌握要求、典型错误。",
    "operationId": "debug_kp_api_debug_kp__kp_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "kp_id",
      "required": true,
      "schema": {
       "title": "Kp Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/KPDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Kp"
   }
  },
...
  "/api/debug/runs": {
   "get": {
    "operationId": "debug_runs_api_debug_runs_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "session_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session Id"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 30,
       "maximum": 100,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Offset",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/DebugRunItem"
         },
         "title": "Response Debug Runs Api Debug Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Runs"
   }
  },
  "/api/debug/runs/{run_id}": {
   "get": {
    "operationId": "debug_run_api_debug_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DebugRunDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Run"
   }
  },
  "/api/debug/runs/{run_id}/events": {
   "get": {
    "operationId": "debug_events_api_debug_runs__run_id__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "after",
      "required": false,
      "schema": {
       "default": 0,
       "title": "After",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "discriminator": {
           "mapping": {
            "checkpoint.requested": "#/components/schemas/CheckpointRequested",
            "item.delivered": "#/components/schemas/ItemDelivered",
            "item.status": "#/components/schemas/ItemStatus",
            "llm.call": "#/components/schemas/LLMCall",
            "message.delta": "#/components/schemas/MessageDelta",
            "message.done": "#/components/schemas/MessageDone",
            "paper.patch": "#/components/schemas/PaperPatched",
            "progress": "#/components/schemas/Progress",
            "retrieval.result": "#/components/schemas/RetrievalResult",
            "run.finished": "#/components/schemas/RunFinished",
            "run.paused": "#/components/schemas/RunPaused",
            "run.started": "#/components/schemas/RunStarted",
            "span.finished": "#/components/schemas/SpanFinished",
            "span.started": "#/components/schemas/SpanStarted",
            "understanding.ready": "#/components/schemas/UnderstandingReady",
            "usage.update": "#/components/schemas/UsageUpdate"
           },
           "propertyName": "type"
          },
          "oneOf": [
           {
            "$ref": "#/components/schemas/RunStarted"
           },
           {
            "$ref": "#/components/schemas/RunFinished"
           },
           {
            "$ref": "#/components/schemas/RunPaused"
           },
           {
            "$ref": "#/components/schemas/SpanStarted"
           },
           {
            "$ref": "#/components/schemas/SpanFinished"
           },
           {
            "$ref": "#/components/schemas/Progress"
           },
           {
            "$ref": "#/components/schemas/MessageDelta"
           },
           {
            "$ref": "#/components/schemas/MessageDone"
           },
           {
            "$ref": "#/components/schemas/ItemStatus"
           },
           {
            "$ref": "#/components/schemas/ItemDelivered"
           },
           {
            "$ref": "#/components/schemas/PaperPatched"
           },
           {
            "$ref": "#/components/schemas/CheckpointRequested"
           },
           {
            "$ref": "#/components/schemas/UnderstandingReady"
           },
           {
            "$ref": "#/components/schemas/LLMCall"
           },
           {
            "$ref": "#/components/schemas/RetrievalResult"
           },
           {
            "$ref": "#/components/schemas/UsageUpdate"
           }
          ]
         },
         "title": "Response Debug Events Api Debug Runs  Run Id  Events Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Events"
   }
  },
  "/api/debug/runs/{run_id}/stream": {
   "get": {
    "description": "调试台的实时事件流：与用户端同一协议，但包含 debug 可见性的事件（span、模型调用、检索）。",
    "operationId": "debug_stream_api_debug_runs__run_id__stream_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "after",
      "required": false,
      "schema": {
       "default": 0,
       "title": "After",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     },
     {
      "in": "header",
      "name": "last-event-id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Last-Event-Id"
      }
     },
     {
      "in": "header",
      "name": "x-debug-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Debug-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Debug Stream"
   }
  },
...
  "/api/knowledge/refs": {
   "get": {
    "description": "知识点的教师可读名称与位置，用于在题目旁显示\"涉及：小数加减法（四下·第一单元）\"。",
    "operationId": "knowledge_refs_api_knowledge_refs_get",
    "parameters": [
     {
      "in": "query",
      "name": "ids",
      "required": true,
      "schema": {
       "items": {
        "type": "string"
       },
       "title": "Ids",
       "type": "array"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/KPRef"
         },
         "title": "Response Knowledge Refs Api Knowledge Refs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Knowledge Refs"
   }
  },
  "/api/runs/{run_id}": {
   "get": {
    "operationId": "get_run_api_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run"
   }
  },
  "/api/runs/{run_id}/cancel": {
   "post": {
    "operationId": "cancel_run_api_runs__run_id__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "boolean"
         },
         "title": "Response Cancel Run Api Runs  Run Id  Cancel Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cancel Run"
   }
  },
  "/api/runs/{run_id}/checkpoint": {
   "post": {
    "operationId": "answer_checkpoint_api_runs__run_id__checkpoint_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CheckpointAnswer"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "boolean"
         },
         "title": "Response Answer Checkpoint Api Runs  Run Id  Checkpoint Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Answer Checkpoint"
   }
  },
  "/api/runs/{run_id}/events": {
   "get": {
    "operationId": "run_events_api_runs__run_id__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "after",
      "required": false,
      "schema": {
       "default": 0,
       "title": "After",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "last-event-id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Last-Event-Id"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run Events"
   }
  },
...
  "/api/sessions/{session_id}": {
   "get": {
    "operationId": "get_session_api_sessions__session_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "session_id",
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SessionState"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Session"
   }
  },
  "/api/sessions/{session_id}/export": {
   "post": {
    "description": "导出当前试卷。`inline=true` 用于页面内预览（PDF 直接嵌入）；警告（缺字、图无法绘制等）放在 `X-Export-Warnings`。",
    "operationId": "export_paper_file_api_sessions__session_id__export_post",
    "parameters": [
     {
      "in": "path",
      "name": "session_id",
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "inline",
      "required": false,
      "schema": {
       "default": false,
       "title": "Inline",
       "type": "boolean"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExportOptions"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/octet-stream": {}
      },
      "description": "导出的文件"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Paper File"
   }
  },
  "/api/sessions/{session_id}/figures/{figure_id}": {
   "get": {
    "description": "试卷里某个图形的 SVG（预览与导出共用同一份渲染）。",
    "operationId": "paper_figure_api_sessions__session_id__figures__figure_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "session_id",
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "figure_id",
      "required": true,
      "schema": {
       "title": "Figure Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "image/svg+xml": {}
      },
      "description": "图形的 SVG"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Paper Figure"
   }
  },
  "/api/sessions/{session_id}/paper": {
   "patch": {
    "description": "手动编辑：应用补丁并立即返回新修订；内容被改的题状态回到\"待核验\"，同时发起一个复核运行（`review_run_id`）。",
    "operationId": "patch_paper_api_sessions__session_id__paper_patch",
    "parameters": [
     {
      "in": "path",
      "name": "session_id",
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PaperPatchBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PaperUpdate"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Paper"
   }
  },
...
 }
}